+++ spec/models/user_session_spec.rb
@@ -1,132 +1,73 @@
-require "rails_helper"
-
 RSpec.describe UserSession, type: :model do
-  include Rails.application.routes.url_helpers
+  subject(:user_session) { described_class.new(session: session) }
+
   let(:session) { instance_double(ActionDispatch::Request::Session) }
 
-  describe "#persist_successful_dfe_sign_in_claim!" do
-    it "stores the DfE Sign-in UID in the session" do
-      omniauth_hash = {
-        "uid" => "123",
-        "credentials" => {
-          "id_token" => "456",
-        },
-      }
-
-      user_session = described_class.new(session: session)
-
-      expect(session).to receive(:[]=).with(:dfe_sign_in_uid, "123")
-      expect(session).to receive(:[]=).with(:dfe_sign_in_sign_out_token, "456")
-
-      user_session.persist_successful_dfe_sign_in_claim!(omniauth_hash: omniauth_hash)
-    end
+  let(:omniauth_hash) do
+    {
+      "uid" => "123",
+      "credentials" => {
+        "id_token" => "456",
+      },
+    }
   end
 
-  describe "#repudiate!" do
-    it "removes the DfE sign in claim from the session" do
-      user_session = described_class.new(session: session)
-
+  describe "#delete!" do
+    it "deletes the DfE sign in claim from the session" do
       expect(session).to receive(:delete).with(:dfe_sign_in_uid)
       expect(session).to receive(:delete).with(:dfe_sign_in_sign_out_token)
-
-      user_session.repudiate!
+      user_session.delete!
     end
   end
 
   describe "#should_be_signed_out_of_dsi?" do
     it "returns true when there is a sign out token" do
-      session_with_sign_out_token = instance_double(ActionDispatch::Request::Session)
-      allow(session_with_sign_out_token)
-        .to receive(:key?)
-        .with(:dfe_sign_in_sign_out_token)
-        .and_return(true)
-
-      allow(session_with_sign_out_token)
-        .to receive(:[])
-        .with(:dfe_sign_in_sign_out_token)
-        .and_return("a-long-token")
-
-      result = described_class.new(session: session_with_sign_out_token)
-        .should_be_signed_out_of_dsi?
-
-      expect(result).to eq(true)
+      allow(session).to receive(:[]).with(:dfe_sign_in_sign_out_token).and_return("a-long-token")
+      expect(user_session.should_be_signed_out_of_dsi?).to be true
     end
 
     it "returns false when the sign out token value is blank" do
-      session_with_sign_out_token = instance_double(ActionDispatch::Request::Session)
-      allow(session_with_sign_out_token)
-        .to receive(:key?)
-        .with(:dfe_sign_in_sign_out_token)
-        .and_return(true)
-
-      allow(session_with_sign_out_token)
-        .to receive(:[])
-        .with(:dfe_sign_in_sign_out_token)
-        .and_return("")
-
-      result = described_class.new(session: session_with_sign_out_token)
-        .should_be_signed_out_of_dsi?
-
-      expect(result).to eq(false)
+      allow(session).to receive(:[]).with(:dfe_sign_in_sign_out_token).and_return("")
+      expect(user_session.should_be_signed_out_of_dsi?).to be false
     end
 
     it "returns false when there is no sign out token" do
-      session_without_sign_in_token = instance_double(ActionDispatch::Request::Session)
-      allow(session_without_sign_in_token)
-        .to receive(:key?)
-        .with(:dfe_sign_in_sign_out_token)
-        .and_return(false)
+      allow(session).to receive(:[]).with(:dfe_sign_in_sign_out_token).and_return(nil)
+      expect(user_session.should_be_signed_out_of_dsi?).to be false
+    end
+  end
 
-      result = described_class.new(session: session_without_sign_in_token)
-        .should_be_signed_out_of_dsi?
-
-      expect(result).to eq(false)
+  describe "#redirect_url" do
+    it "returns the application root uri" do
+      expect(user_session.redirect_url).to eql "http://localhost:3000/"
     end
   end
 
   describe "#sign_out_url" do
     around do |example|
-      ClimateControl.modify(
-        DFE_SIGN_IN_ISSUER: "https://test-oidc.signin.education.gov.uk:443",
-      ) do
+      ClimateControl.modify(DFE_SIGN_IN_ISSUER: "https://test-oidc.signin.education.gov.uk:443") do
         example.run
       end
     end
 
     it "returns a URL that can be sent to DfE Sign-in to sign the user out of their service" do
-      session_with_sign_out_token = instance_double(ActionDispatch::Request::Session)
-
-      allow(session_with_sign_out_token)
-        .to receive(:key?)
-        .with(:dfe_sign_in_sign_out_token)
-        .and_return(true)
-
-      allow(session_with_sign_out_token)
-        .to receive(:[])
-        .with(:dfe_sign_in_sign_out_token)
-        .and_return("a-long-token")
-
-      user_session = described_class.new(session: session_with_sign_out_token)
-
-      result = user_session.sign_out_url
-
-      expect(result).to eq("https://test-oidc.signin.education.gov.uk:443/session/end?id_token_hint=a-long-token&post_logout_redirect_uri=http%3A%2F%2Flocalhost%3A3000%2Fauth%2Fdfe%2Fsignout")
+      allow(session).to receive(:[]).with(:dfe_sign_in_sign_out_token).and_return("a-long-token")
+      expect(user_session.sign_out_url).to eq("https://test-oidc.signin.education.gov.uk:443/session/end?id_token_hint=a-long-token&post_logout_redirect_uri=http%3A%2F%2Flocalhost%3A3000%2F")
     end
 
     context "when the user has no sign out token" do
-      it "redirects the user to the root path" do
-        session_without_sign_in_token = instance_double(ActionDispatch::Request::Session)
-        allow(session_without_sign_in_token)
-          .to receive(:key?)
-          .with(:dfe_sign_in_sign_out_token)
-          .and_return(false)
-
-        user_session = described_class.new(session: session_without_sign_in_token)
-
-        result = user_session.sign_out_url
-
-        expect(result).to eql(root_path)
+      it "redirects the user to the root uri" do
+        allow(session).to receive(:[]).with(:dfe_sign_in_sign_out_token).and_return(nil)
+        expect(user_session.sign_out_url).to eql "http://localhost:3000/"
       end
+    end
+  end
+
+  describe "#persist_successful_dfe_sign_in_claim!" do
+    it "stores the DfE Sign-in UID in the session" do
+      expect(session).to receive(:[]=).with(:dfe_sign_in_uid, "123")
+      expect(session).to receive(:[]=).with(:dfe_sign_in_sign_out_token, "456")
+      user_session.persist_successful_dfe_sign_in_claim!(omniauth_hash: omniauth_hash)
     end
   end
 
@@ -141,19 +82,8 @@ RSpec.describe UserSession, type: :model do
         instance_double(Rack::Session::SessionId, private_id: "2::5347845262539"))
     end
 
-    let(:omniauth_hash) do
-      {
-        "uid" => "123",
-        "credentials" => {
-          "id_token" => "456",
-        },
-      }
-    end
-
     context "when no other session exists for that DfE Sign in user" do
       it "does not delete any pre-existing session data from the Redis session store" do
-        user_session = described_class.new(session: session)
-
         fake_session_redis = MockRedis.new
         allow(RedisSessions).to receive(:redis).and_return(fake_session_redis)
         expect(fake_session_redis).not_to receive(:del)
@@ -162,8 +92,6 @@ RSpec.describe UserSession, type: :model do
       end
 
       it "adds a new session lookup" do
-        user_session = described_class.new(session: session)
-
         fake_session_lookup_redis = MockRedis.new
         allow(RedisSessionLookup).to receive(:redis).and_return(fake_session_lookup_redis)
         expect(fake_session_lookup_redis).to receive(:set).with("user_dsi_id:123", "2::5347845262539")
@@ -177,8 +105,6 @@ RSpec.describe UserSession, type: :model do
         RedisSessions.redis.set("session:2::5347845262539",
                                 Marshal.dump({ "_csrf_token" => "1", "dfe_sign_in_uid" => "123" }))
         RedisSessionLookup.redis.set("user_dsi_id:123", "2::5347845262539")
-
-        user_session = described_class.new(session: session)
 
         fake_session_redis = MockRedis.new
         allow(RedisSessions).to receive(:redis).and_return(fake_session_redis)